reorganises form input field components into separate file
